+++ src/pydvl/parallel/futures/ray.py
@@ -97,7 +97,7 @@ class RayExecutor(Executor):
         # Work Item Manager Thread
         self._work_item_manager_thread: Optional[_WorkItemManagerThread] = None
 
-    def submit(self, fn: Callable[..., T], *args, **kwargs) -> "Future[T]":
+    def submit(self, fn: Callable[..., T], /, *args, **kwargs) -> Future[T]:
         r"""Submits a callable to be executed with the given arguments.
 
         Schedules the callable to be executed as fn(\*args, \**kwargs)
@@ -240,7 +240,7 @@ class _WorkItem:
         self.future.object_ref = ref  # type: ignore
 
     if sys.version_info >= (3, 9):
-        __class_getitem__ = classmethod(types.GenericAlias)
+        __class_getitem__ = classmethod(types.GenericAlias)  # type: ignore
 
 
 class _WorkItemManagerThread(threading.Thread):