+++ src/pydvl/valuation/dataset.py
@@ -1,48 +1,52 @@
 """
 This module contains convenience classes to handle data and groups thereof.
 
-Model-based value computations require evaluation of a scoring function (the *utility*).
-This is typically the performance of the model on a test set (as an approximation to its
-true expected performance). It is therefore convenient to keep both the training data
-and the test data grouped to be passed around to methods in [shapley][pydvl.valuation].
-This is done with [Dataset][pydvl.valuation.dataset.Dataset].
+Value computations with supervised models benefit from a unified interface to handle
+data. This module provides two classes to handle data and labels, as well as feature
+names and other information:
+[Dataset][pydvl.valuation.dataset.Dataset]
+and [GroupedDataset][pydvl.valuation.dataset.GroupedDataset]. Objects of both types can
+be used to construct [scorers][pydvl.valuation.scorers] and to fit (most) valuation
+methods.
 
-These underlying data arrays can be accessed via
-[Dataset.data][pydvl.valuation.dataset.Dataset.data], which returns the tuple `(X, y)`.
-The data can be accessed by indexing the object directly, e.g. `dataset[0]` will return
-the data point corresponding to index 0 in `dataset`. Note however that this is not
-necessarily the same as `dataset.data().x[0]`, which is the first point in the data
-array. This is in particular true for
+The underlying data arrays can always be accessed via
+[Dataset.data][pydvl.valuation.dataset.Dataset.data], which returns the tuple `(x, y)`.
+
+Slicing the object, e.g. `dataset[0]`, will return a new `Dataset` with the data
+corresponding to that slice. Note however that the contents of the new object, i.e.
+`dataset[0].data().x`, may not be the same as `dataset.data().x[0]`, which is the first
+point in the original data array. This is in particular true for
 [GroupedDatasets][pydvl.valuation.dataset.GroupedDataset] where one "logical" index may
-correspond to multiple data points. 
+correspond to multiple data points.
 
-Objects of both types can be used to construct [scorers][pydvl.valuation.scorers] and to
-fit (most) valuation methods.
 
 ## Grouped datasets and logical indices
 
-It is also possible to group data points together with
+As mentioned above, it is also possible to group data points together with
 [GroupedDataset][pydvl.valuation.dataset.dataset.GroupedDataset].
-
-A call to [Dataset.data(indices)][pydvl.valuation.dataset.Dataset.data] will return the
-data and labels of all samples for the given groups. But `grouped_data[0]` will return
-the data and labels of the first group, not the first data point and will therefore be
-in general different from `grouped_data.data([0])`.
-
 In order to handle groups correctly, Datasets map "logical" indices to "data" indices
 and vice versa. The latter correspond to indices in the data arrays themselves, while
 the former may map to groups of data points.
 
-This is important for valuation methods that require computation on individual data
-points, like KNNShapley or Data-OOB. In these cases, the logical indices are used to
-compute the Shapley values, while the data indices are used internally by the method.
+A call to [GroupedDataset.data(indices)][pydvl.valuation.dataset.GroupedDataset.data]
+will return the data and labels of all samples for the given groups. But
+`grouped_data[0]` will return the data and labels of the first group, not the first data
+point and will therefore be in general different from `grouped_data.data([0])`.
+
+Grouping data can be useful to reduce computation time, e.g. for Shapley-based methods.
+
+It is important to keep in mind the distinction between logical and data indices for
+valuation methods that require computation on individual data points, like KNNShapley or
+Data-OOB. In these cases, the logical indices are used to compute the Shapley values,
+while the data indices are used internally by the method.
 """
 
 from __future__ import annotations
 
 import logging
 from collections import OrderedDict
-from typing import Any, Iterable, NamedTuple, Sequence
+from dataclasses import dataclass
+from typing import Sequence, overload
 
 import numpy as np
 from numpy.typing import NDArray
@@ -55,9 +59,14 @@ __all__ = ["Dataset", "GroupedDataset", "RawData"]
 logger = logging.getLogger(__name__)
 
 
-class RawData(NamedTuple):
+@dataclass(frozen=True)
+class RawData:
     x: NDArray
     y: NDArray
+
+    # Make the unpacking operator work
+    def __iter__(self):  # No way to type the return Iterator properly
+        return iter((self.x, self.y))
 
 
 class Dataset:
@@ -76,15 +85,17 @@ class Dataset:
     """
 
     _indices: NDArray[np.int_]
-    _data_names: NDArray[np.object_]
+    _data_names: NDArray[np.str_]
+    feature_names: list[str]
+    target_names: list[str]
 
     def __init__(
         self,
         x: NDArray,
         y: NDArray,
-        feature_names: Sequence[str] | None = None,
-        target_names: Sequence[str] | None = None,
-        data_names: Sequence[str] | None = None,
+        feature_names: Sequence[str] | NDArray[np.str_] | None = None,
+        target_names: Sequence[str] | NDArray[np.str_] | None = None,
+        data_names: Sequence[str] | NDArray[np.str_] | None = None,
         description: str | None = None,
         multi_output: bool = False,
     ):
@@ -105,6 +116,8 @@ class Dataset:
 
         !!! tip "Changed in version 0.10.0"
             No longer holds split data, but only x, y.
+        !!! tip "Changed in version 0.10.0"
+            Slicing now return a new `Dataset` object, not raw data.
         """
         self._x, self._y = check_X_y(
             x, y, multi_output=multi_output, estimator="Dataset"
@@ -114,14 +127,12 @@ class Dataset:
             n = a.shape[1] if len(a.shape) > 1 else 1
             return [f"{s}{i:0{1 + int(np.log10(n))}d}" for i in range(1, n + 1)]
 
-        self.feature_names = feature_names
-        self.target_names = target_names
-
-        if self.feature_names is None:
-            self.feature_names = make_names("x", x)
-
-        if self.target_names is None:
-            self.target_names = make_names("y", y)
+        self.feature_names = (
+            list(feature_names) if feature_names is not None else make_names("x", x)
+        )
+        self.target_names = (
+            list(target_names) if target_names is not None else make_names("y", y)
+        )
 
         if len(self._x.shape) > 1:
             if len(self.feature_names) != self._x.shape[-1]:
@@ -133,12 +144,12 @@ class Dataset:
         self.description = description or "No description"
         self._indices = np.arange(len(self._x), dtype=np.int_)
         self._data_names = (
-            np.array(data_names, dtype=object)
+            np.array(data_names, dtype=np.str_)
             if data_names is not None
-            else self._indices.astype(object)
+            else self._indices.astype(np.str_)
         )
 
-    def __getitem__(self, idx: int | slice | Iterable) -> Dataset:
+    def __getitem__(self, idx: int | slice | Sequence[int]) -> Dataset:
         if isinstance(idx, int):
             idx = [idx]
         return Dataset(
@@ -152,11 +163,13 @@ class Dataset:
 
     def feature(self, name: str) -> tuple[slice, int]:
         try:
-            return np.index_exp[:, self.feature_names.index(name)]
+            return np.index_exp[:, self.feature_names.index(name)]  # type: ignore
         except ValueError:
             raise ValueError(f"Feature {name} is not in {self.feature_names}")
 
-    def data(self, indices: Iterable[int] | None = None) -> RawData:
+    def data(
+        self, indices: int | slice | Sequence[int] | NDArray[np.int_] | None = None
+    ) -> RawData:
         """Given a set of indices, returns the training data that refer to those
         indices.
 
@@ -177,7 +190,7 @@ class Dataset:
             return RawData(self._x, self._y)
         return RawData(self._x[indices], self._y[indices])
 
-    def data_indices(self, indices: Iterable[int] | None = None) -> NDArray[np.int_]:
+    def data_indices(self, indices: Sequence[int] | None = None) -> NDArray[np.int_]:
         """Returns a subset of indices.
 
         This is equivalent to using `Dataset.indices[logical_indices]` but allows
@@ -196,8 +209,8 @@ class Dataset:
             return self._indices
         return self._indices[indices]
 
-    def logical_indices(self, indices: Iterable[int] | None = None) -> NDArray[np.int_]:
-        """Returns the indices in this Dataset for the given indices in the data array.
+    def logical_indices(self, indices: Sequence[int] | None = None) -> NDArray[np.int_]:
+        """Returns the indices in this `Dataset` for the given indices in the data array.
 
         This is equivalent to using `Dataset.indices[data_indices]` but allows
         subclasses to define special behaviour, e.g. when indices in `Dataset` do not
@@ -228,7 +241,7 @@ class Dataset:
         return self._indices
 
     @property
-    def names(self) -> NDArray[np.object_]:
+    def names(self) -> NDArray[np.str_]:
         """Names of each individual datapoint.
 
         Used for reporting Shapley values.
@@ -371,15 +384,17 @@ class Dataset:
 
 
 class GroupedDataset(Dataset):
+    _group_names: NDArray[np.str_]
+
     def __init__(
         self,
         x: NDArray,
         y: NDArray,
-        data_groups: Sequence[int],
-        feature_names: Sequence[str] | None = None,
-        target_names: Sequence[str] | None = None,
-        data_names: Sequence[str] | None = None,
-        group_names: Sequence[str] | None = None,
+        data_groups: Sequence[int] | NDArray[np.int_],
+        feature_names: Sequence[str] | NDArray[np.str_] | None = None,
+        target_names: Sequence[str] | NDArray[np.str_] | None = None,
+        data_names: Sequence[str] | NDArray[np.str_] | None = None,
+        group_names: Sequence[str] | NDArray[np.str_] | None = None,
         description: str | None = None,
         **kwargs,
     ):
@@ -392,7 +407,7 @@ class GroupedDataset(Dataset):
         Args:
             x: training data
             y: labels of training data
-            data_groups: Iterable of the same length as `x_train` containing
+            data_groups: Sequence of the same length as `x_train` containing
                 a group id for each training data point. Data points with the same
                 id will then be grouped by this object and considered as one for
                 effects of valuation. Group ids are assumed to be zero-based consecutive
@@ -431,24 +446,24 @@ class GroupedDataset(Dataset):
             )
 
         # data index -> abstract index (group id)
-        self.data_to_group = np.array(data_groups, dtype=int)
+        self.data_to_group: NDArray[np.int_] = np.array(data_groups, dtype=int)
         # abstract index (group id) -> data index
         self.group_to_data: OrderedDict[int, list[int]] = OrderedDict(
             {k: [] for k in set(data_groups)}
         )
         for data_idx, group_idx in enumerate(self.data_to_group):
             self.group_to_data[group_idx].append(data_idx)  # type: ignore
-        self._indices = np.array(list(self.group_to_data.keys()))
+        self._indices = np.array(list(self.group_to_data.keys()), dtype=np.int_)
         self._group_names = (
-            np.array(group_names, dtype=object)
+            np.array(group_names, dtype=np.str_)
             if group_names is not None
-            else np.array(list(self.group_to_data.keys()), dtype=object)
+            else np.array(list(self.group_to_data.keys()), dtype=np.str_)
         )
 
     def __len__(self):
         return len(self._indices)
 
-    def __getitem__(self, idx: int | slice | Iterable) -> GroupedDataset:
+    def __getitem__(self, idx: int | slice | Sequence[int]) -> GroupedDataset:
         if isinstance(idx, int):
             idx = [idx]
         return GroupedDataset(
@@ -468,12 +483,14 @@ class GroupedDataset(Dataset):
         return self._indices
 
     @property
-    def names(self) -> NDArray[object]:
+    def names(self) -> NDArray[np.str_]:
         """Names of the groups."""
         # FIXME? this shadows _data_names (but it can still be accessed...)
         return self._group_names
 
-    def data(self, indices: Iterable[int] | None = None) -> RawData:
+    def data(
+        self, indices: int | slice | Sequence[int] | NDArray[np.int_] | None = None
+    ) -> RawData:
         """Returns the data and labels of all samples in the given groups.
 
         Args:
@@ -485,7 +502,9 @@ class GroupedDataset(Dataset):
         """
         return super().data(self.data_indices(indices))
 
-    def data_indices(self, indices: Iterable[int] | None = None) -> NDArray[np.int_]:
+    def data_indices(
+        self, indices: int | slice | Sequence[int] | NDArray[np.int_] | None = None
+    ) -> NDArray[np.int_]:
         """Returns the indices of the samples in the given groups.
 
         Args:
@@ -501,7 +520,7 @@ class GroupedDataset(Dataset):
             indices = range(*indices.indices(len(self.group_to_data)))
         return np.concatenate([self.group_to_data[i] for i in indices], dtype=np.int_)  # type: ignore
 
-    def logical_indices(self, indices: Iterable[int] | None = None) -> NDArray[np.int_]:
+    def logical_indices(self, indices: Sequence[int] | None = None) -> NDArray[np.int_]:
         """Returns the group indices for the given data indices.
 
         Args:
@@ -515,14 +534,37 @@ class GroupedDataset(Dataset):
             return self.data_to_group
         return self.data_to_group[indices]
 
+    @overload
     @classmethod
     def from_sklearn(
         cls,
         data: Bunch,
         train_size: float = 0.8,
-        stratify_by_target: bool = False,
-        data_groups: Sequence | None = None,
         random_state: int | None = None,
+        stratify_by_target: bool = False,
+        **kwargs,
+    ) -> tuple[GroupedDataset, GroupedDataset]: ...
+
+    @overload
+    @classmethod
+    def from_sklearn(
+        cls,
+        data: Bunch,
+        train_size: float = 0.8,
+        random_state: int | None = None,
+        stratify_by_target: bool = False,
+        data_groups: Sequence[int] | None = None,
+        **kwargs,
+    ) -> tuple[GroupedDataset, GroupedDataset]: ...
+
+    @classmethod
+    def from_sklearn(
+        cls,
+        data: Bunch,
+        train_size: float = 0.8,
+        random_state: int | None = None,
+        stratify_by_target: bool = False,
+        data_groups: Sequence[int] | None = None,
         **kwargs,
     ) -> tuple[GroupedDataset, GroupedDataset]:
         """Constructs a [GroupedDataset][pydvl.valuation.dataset.GroupedDataset] object, and an
@@ -542,7 +584,6 @@ class GroupedDataset(Dataset):
 
         Args:
             data: scikit-learn Bunch object. The following attributes are supported:
-
                 - `data`: covariates.
                 - `target`: target variables (labels).
                 - `feature_names` (**optional**): the feature names.
@@ -566,15 +607,41 @@ class GroupedDataset(Dataset):
             Returns a tuple of two [GroupedDataset][pydvl.valuation.dataset.GroupedDataset]
                 objects.
         """
+
         return cls.from_arrays(
             X=data.data,
             y=data.target,
             train_size=train_size,
+            random_state=random_state,
             stratify_by_target=stratify_by_target,
             data_groups=data_groups,
-            random_state=random_state,
             **kwargs,
         )
+
+    @overload
+    @classmethod
+    def from_arrays(
+        cls,
+        X: NDArray,
+        y: NDArray,
+        train_size: float = 0.8,
+        random_state: int | None = None,
+        stratify_by_target: bool = False,
+        **kwargs,
+    ) -> tuple[GroupedDataset, GroupedDataset]: ...
+
+    @overload
+    @classmethod
+    def from_arrays(
+        cls,
+        X: NDArray,
+        y: NDArray,
+        train_size: float = 0.8,
+        random_state: int | None = None,
+        stratify_by_target: bool = False,
+        data_groups: Sequence[int] | None = None,
+        **kwargs,
+    ) -> tuple[GroupedDataset, GroupedDataset]: ...
 
     @classmethod
     def from_arrays(
@@ -582,9 +649,9 @@ class GroupedDataset(Dataset):
         X: NDArray,
         y: NDArray,
         train_size: float = 0.8,
-        stratify_by_target: bool = False,
-        data_groups: Sequence | None = None,
         random_state: int | None = None,
+        stratify_by_target: bool = False,
+        data_groups: Sequence[int] | None = None,
         **kwargs,
     ) -> tuple[GroupedDataset, GroupedDataset]:
         """Constructs a [GroupedDataset][pydvl.valuation.dataset.GroupedDataset] object,
@@ -636,6 +703,7 @@ class GroupedDataset(Dataset):
             Returns a tuple of two [GroupedDataset][pydvl.valuation.dataset.GroupedDataset]
                 objects.
         """
+
         if data_groups is None:
             raise ValueError(
                 "data_groups must be provided when constructing a GroupedDataset"
@@ -653,7 +721,7 @@ class GroupedDataset(Dataset):
         return training_set, test_set
 
     @classmethod
-    def from_dataset(cls, data: Dataset, data_groups: Sequence[Any]) -> GroupedDataset:
+    def from_dataset(cls, data: Dataset, data_groups: Sequence[int]) -> GroupedDataset:
         """Creates a [GroupedDataset][pydvl.valuation.dataset.GroupedDataset] object from a
         [Dataset][pydvl.valuation.dataset.Dataset] object and a mapping of data groups.